Moving old dhnsw code temporarily and setting up space for dhnsw code with new API
